feat: Add a mock to test an external API
Though this code needs to be refactored and cleaned, I added it anyway as a commit to use it as a middle step towards learning how to write tests (and what is a mock).
A `utils` folder was added in `src`, with a function to fetch data from the API. The fetch function was added to the `componentDidMount` method of the User component to update its state. A `mocks` folder was added to `utils` to use the `jest.mock` API.
The `App` test was deleted.