refactor: Correct use of `match` and path.
Deleted wrong code inside Route, where an array containing some hardcoded text had been added to test the Route.
The issue in that that array was not what should be expected as output of the `matchPath` function. It is instead a result array that could come as output of the `exec` method of a regular expression.