feat: Add Flow flag to JavaScript files
Added the `@flow` flag to JavaScript files to let Flow know which files to monitor and check.
Added the `flow-typed` npm package to provide a library directory for third-party libraries used by the project. This lets Flow treat them as untyped dependencies.

IMPORTANT: as the `flow-typed` cli looks into the `devDependencies` property in the `package.json` file, Jest needed to be added there to be found by the `flow-typed` cli.
That's because Jest causes Flow to warn against its unknown syntax and this project was created by `create-react-app` which uses Jest but does not list it as a dev dependency.