Fix: fix bug
Added the `exact` prop to the Route to a user item to address a problem that occurred when two users had an id that started with the same number.
In this case, before adding this prop, when clicking on the link of the user with id = 10, both this user and user with id = 1 were displayed on the page.